+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - fixed: Disabling pin-login while in duress mode disables pin-login for the main login.
 - fixed: `loginWithPassword` bug disabled duress mode when doing an online login.
+- fixed: Properly fake duress mode settings while in duress mode.
 
 ## 2.27.4 (2025-05-13)